// types.ts
export interface ShadowObject {
  text: string;
  x: number;
  y: number;
  width: number;
  height: number;
}